Move presenters into govspeak namespace
This resolves conflicts we have where codebases use classes such as
AttachmentPresenter.

Also fixes a few coding standards issues.
