+++ lib/govspeak.rb
@@ -1,6 +1,9 @@
-require 'kramdown'
 require 'active_support/core_ext/hash'
 require 'active_support/core_ext/array'
+require 'erb'
+require 'htmlentities'
+require 'kramdown'
+require 'kramdown/parser/kramdown_with_automatic_external_links'
 require 'govspeak/header_extractor'
 require 'govspeak/structured_header_extractor'
 require 'govspeak/html_validator'
@@ -8,12 +11,9 @@ require 'govspeak/html_sanitizer'
 require 'govspeak/kramdown_overrides'
 require 'govspeak/blockquote_extra_quote_remover'
 require 'govspeak/post_processor'
-require 'kramdown/parser/kramdown_with_automatic_external_links'
-require 'htmlentities'
-require 'presenters/attachment_presenter'
-require 'presenters/contact_presenter'
-require 'presenters/h_card_presenter'
-require 'erb'
+require 'govspeak/presenters/attachment_presenter'
+require 'govspeak/presenters/contact_presenter'
+require 'govspeak/presenters/h_card_presenter'
 
 module Govspeak
